flask/requirements.txt
bidict==0.22.1
click==8.1.6
coverage==7.3.2
dnspython==2.4.2
eventlet==0.33.3
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
peewee==3.16.3
python-engineio==4.8.0
python-socketio==5.9.0
six==1.16.0
Werkzeug==2.3.8
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
# Database drivers for external databases
psycopg2-binary==2.9.7
PyMySQL==1.1.0
# Environment variables support
python-dotenv==1.0.0
